Separate the Carpaccio HTTP worker and the Carpaccio server in different files.
